docs: documentation Service and Controller, little fix.
